chore: cleaning unused onboarding screen and associated components
